Move split_entity_id to helpers
